requirements.txt
playwright==1.40.0
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.0

